+++ soccer/CMakeLists.txt
@@ -26,13 +26,13 @@ set(ROBOCUP_LIB_SRC
     "optimization/ParallelGradientAscent1D.cpp"
     "optimization/NelderMead2D.cpp"
     "optimization/PythonFunctionWrapper.cpp"
-    "planning/low_level/AnglePlanning.cpp"
-    "planning/low_level/CreatePath.cpp"
-    "planning/low_level/PathSmoothing.cpp"
-    "planning/low_level/Replanner.cpp"
-    "planning/low_level/RRTUtil.cpp"
-    "planning/low_level/TrapezoidalMotion.cpp"
-    "planning/low_level/VelocityProfiling.cpp"
+    "planning/primitives/AnglePlanning.cpp"
+    "planning/primitives/CreatePath.cpp"
+    "planning/primitives/PathSmoothing.cpp"
+    "planning/primitives/Replanner.cpp"
+    "planning/primitives/RRTUtil.cpp"
+    "planning/primitives/TrapezoidalMotion.cpp"
+    "planning/primitives/VelocityProfiling.cpp"
     "planning/planner/CollectPlanner.cpp"
     "planning/planner/EscapeObstaclesPathPlanner.cpp"
     "planning/planner/InterceptPlanner.cpp"
