+++ soccer/planning/planner/EscapeObstaclesPathPlanner.cpp
@@ -4,10 +4,10 @@
 #include <optional>
 #include <vector>
 
-#include "planning/low_level/AnglePlanning.hpp"
-#include "planning/low_level/CreatePath.hpp"
-#include "planning/low_level/RRTUtil.hpp"
-#include "planning/low_level/RoboCupStateSpace.hpp"
+#include "planning/primitives/AnglePlanning.hpp"
+#include "planning/primitives/CreatePath.hpp"
+#include "planning/primitives/RRTUtil.hpp"
+#include "planning/primitives/RoboCupStateSpace.hpp"
 
 using namespace Geometry2d;
 namespace Planning {
