+++ soccer/planning/planner/InterceptPlanner.cpp
@@ -4,10 +4,10 @@
 #include <Constants.hpp>
 
 #include "planning/Instant.hpp"
-#include "planning/low_level/AnglePlanning.hpp"
-#include "planning/low_level/CreatePath.hpp"
-#include "planning/low_level/RRTUtil.hpp"
-#include "planning/low_level/VelocityProfiling.hpp"
+#include "planning/primitives/AnglePlanning.hpp"
+#include "planning/primitives/CreatePath.hpp"
+#include "planning/primitives/RRTUtil.hpp"
+#include "planning/primitives/VelocityProfiling.hpp"
 
 namespace Planning {
