+++ soccer/planning/primitives/CreatePath.cpp
@@ -1,8 +1,8 @@
 #include "CreatePath.hpp"
 
 #include "planning/TrajectoryUtils.hpp"
-#include "planning/low_level/RRTUtil.hpp"
-#include "planning/low_level/VelocityProfiling.hpp"
+#include "planning/primitives/RRTUtil.hpp"
+#include "planning/primitives/VelocityProfiling.hpp"
 
 using namespace Geometry2d;
