+++ soccer/planning/primitives/Replanner.cpp
@@ -5,9 +5,9 @@
 #include "RRTUtil.hpp"
 #include "planning/Instant.hpp"
 #include "planning/TrajectoryUtils.hpp"
-#include "planning/low_level/AnglePlanning.hpp"
-#include "planning/low_level/CreatePath.hpp"
 #include "planning/planner/Planner.hpp"
+#include "planning/primitives/AnglePlanning.hpp"
+#include "planning/primitives/CreatePath.hpp"
 
 using namespace Geometry2d;
